requirements.txt
accelerate
av
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
decord
diffusers==0.28
einops
gradio
gradio_client
huggingface_hub==0.25.2
hydra-core
imageio
imageio-ffmpeg
numpy==1.23.5
omegaconf
onnxruntime-gpu
open-clip-torch
opencv-contrib-python
opencv-python
Pillow
scikit-image
scikit-learn
scipy
torch
torchdiffeq
torchmetrics
torchvision
tqdm
transformers
mlflow
xformers==0.0.27
controlnet-aux
bitsandbytes
imageio
ultralytics
mediapipe
iopath